requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
python-dotenv==1.0.0
google-search-results==2.4.2
openai==1.12.0